Add acceptance test suite, documentation
In order to test both the HTML files content and the naviagation,
the full images gallery is generated from a sample XML souce file.

The generated set of data is them served by a Capybara::Server instance.

Note that once the data was generated (in the `spec/tmp` temporary
directory) you can start a web server on your own to browse
the sample pictures!

Usage:

    rackup -p 4567 # then visit http://localhost:4567