Add containerized server
Add an Nginx server to serve the generated images gallery.

The images gallery is generated by gonzalobulnes/images_gallery,
another image called gonzalobulnes/images_gallery_glue performs
a copy of the files to a volume, which is served statically by
gonzalobulnes/images_gallery_server.

That's probably not an optimal setup, suggestions are welcome!

Fixes #16

Thanks https://stackoverflow.com/a/32509136 and
https://training.docker.com/self-paced-training